Revert "Fixes NPE described in issue 123 http://code.google.com/p/gitblit/issues/detail?id=123"
This reverts commit 344b5e5fa40bf848594f2cadbb7aeeca0dc6877b.

From: james.moger@gitblit.com
Subject: Re: patch
Date: Sun, 19 Aug 2012 22:03:59 -0400

...

I merged your build script fixes for proxied connections.
I committed a change to the GitFilter that preserves the spirit of your
fix, but the code is slightly different.

The nullpointer you discovered (no defined action) while triggerable
from curl/wget/etc is not a real-life scenario from a git client (the
consumer of the /git/ servlet).  Nonetheless it is a null pointer and we
guard against those when appropriate.  The repository model does not
need a null check in the GitFilter because the calling class
(AccessRestrictionFilter) already performs this for us.